requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pyyaml
mlflow
scikit-learn
pymongo[srv]
dill
imbalanced-learn

#-e .